edited id type

+++ Restaurant Chain Management System/Part A/Header Files/Restaurant.h
@@ -25,7 +25,7 @@ enum EntryState { ACTIVE, DELETED, EMPTY };
 class Restaurant 
 {
 private:
-    long long int restaurantId;
+     int restaurantId;
     int numOfEmployees;
     RestaurantType restaurantType;
     string restaurantName;
@@ -40,9 +40,9 @@ public:
     // default constructor
     Restaurant();
     // constructor with parameters
-    Restaurant(long long int, int, RestaurantType, const string &, string[], const CostsVec &, const Cuisine[]);
+    Restaurant(int, int, RestaurantType, const string &, string[], const CostsVec &, const Cuisine[]);
     // constructor to initialize the location array and cuisines array elements one by one
-    Restaurant(long long int, int, RestaurantType, const string &, string, string, string, const CostsVec &,
+    Restaurant(int, int, RestaurantType, const string &, string, string, string, const CostsVec &,
                Cuisine, Cuisine, Cuisine, Cuisine, Cuisine);
     // copy constructor
     Restaurant(const Restaurant &);
@@ -50,7 +50,7 @@ public:
     Restaurant(Restaurant &&);
 
     // Getters
-    long long int getRestaurantId() const;
+     int getRestaurantId() const;
     string getRestaurantName() const;
     int getNumOfEmployees() const;
     RestaurantType getRestaurantType() const;
@@ -63,7 +63,7 @@ public:
     const Cuisine *getAllCuisines() const;
 
     //setters
-    void setRestaurantId(long long int id);
+    void setRestaurantId( int id);
     void setRestaurantName(const string &name);
     void setNumOfEmployees(int employees);
     void setRestaurantType(RestaurantType type);
